+++ src/test/scala/edu/ie3/simona/model/thermal/ThermalGridWithStorageOnlySpec.scala
@@ -80,14 +80,44 @@ class ThermalGridWithStorageOnlySpec
       "deliver the capabilities of the storage" in {
         val tick = 10800 // after three hours
 
-        val gridDemand = thermalGrid.energyDemand(
-          tick,
-          testGridAmbientTemperature,
-          ThermalGrid.startingState(thermalGrid),
-        )
+        val (houseDemand, storageDemand, updatedThermalGridState) =
+          thermalGrid.energyDemandAndUpdatedState(
+            tick,
+            testGridAmbientTemperature,
+            ThermalGrid.startingState(thermalGrid),
+          )
 
-        gridDemand.required should approximate(MegawattHours(0d))
-        gridDemand.possible should approximate(MegawattHours(0.92d))
+        houseDemand.required should approximate(KilowattHours(0d))
+        houseDemand.possible should approximate(KilowattHours(0d))
+        storageDemand.required should approximate(KilowattHours(1150d))
+        storageDemand.possible should approximate(KilowattHours(1150d))
+        updatedThermalGridState.houseState shouldBe None
+        updatedThermalGridState.storageState shouldBe Some(
+          ThermalStorageState(10800, KilowattHours(0d), Kilowatts(0d))
+        )
+      }
+
+      "deliver the capabilities of a half full storage" in {
+        val tick = 10800 // after three hours
+
+        val (houseDemand, storageDemand, updatedThermalGridState) =
+          thermalGrid.energyDemandAndUpdatedState(
+            tick,
+            testGridAmbientTemperature,
+            ThermalGridState(
+              None,
+              Some(ThermalStorageState(0L, KilowattHours(575d), Kilowatts(0d))),
+            ),
+          )
+
+        houseDemand.required should approximate(KilowattHours(0d))
+        houseDemand.possible should approximate(KilowattHours(0d))
+        storageDemand.required should approximate(KilowattHours(0d))
+        storageDemand.possible should approximate(KilowattHours(575d))
+        updatedThermalGridState.houseState shouldBe None
+        updatedThermalGridState.storageState shouldBe Some(
+          ThermalStorageState(10800L, KilowattHours(575d), Kilowatts(0d))
+        )
       }
     }
 
@@ -105,7 +135,7 @@ class ThermalGridWithStorageOnlySpec
             Some(
               ThermalStorageState(
                 0L,
-                KilowattHours(430d),
+                KilowattHours(200d),
                 Kilowatts(0d),
               )
             )
@@ -125,7 +155,7 @@ class ThermalGridWithStorageOnlySpec
                 Some(ThermalStorageState(tick, storedEnergy, qDot)),
               ) =>
             tick shouldBe 0L
-            storedEnergy should approximate(KilowattHours(430d))
+            storedEnergy should approximate(KilowattHours(200d))
             qDot should approximate(testGridQDotConsumptionHigh)
           case _ => fail("Thermal grid state has been calculated wrong.")
         }
@@ -149,6 +179,8 @@ class ThermalGridWithStorageOnlySpec
             testGridAmbientTemperature,
             gridState,
             testGridQDotInfeed,
+            noThermalDemand,
+            thermalDemand,
           )
 
         updatedGridState match {
@@ -157,11 +189,11 @@ class ThermalGridWithStorageOnlySpec
                 Some(ThermalStorageState(tick, storedEnergy, qDot)),
               ) =>
             tick shouldBe 0L
-            storedEnergy should approximate(KilowattHours(230d))
+            storedEnergy should approximate(KilowattHours(0d))
             qDot should approximate(testGridQDotInfeed)
           case _ => fail("Thermal grid state has been calculated wrong.")
         }
-        reachedThreshold shouldBe Some(StorageFull(220800L))
+        reachedThreshold shouldBe Some(StorageFull(276000L))
       }
     }
 
@@ -172,9 +204,11 @@ class ThermalGridWithStorageOnlySpec
           ThermalGrid.startingState(thermalGrid),
           testGridAmbientTemperature,
           testGridQDotInfeed,
+          noThermalDemand,
+          thermalDemand,
         )
 
-        nextThreshold shouldBe Some(StorageFull(220800L))
+        nextThreshold shouldBe Some(StorageFull(276000L))
 
         updatedState match {
           case ThermalGridState(
@@ -182,7 +216,7 @@ class ThermalGridWithStorageOnlySpec
                 Some(ThermalStorageState(tick, storedEnergy, qDot)),
               ) =>
             tick shouldBe 0L
-            storedEnergy should approximate(KilowattHours(230d))
+            storedEnergy should approximate(KilowattHours(0d))
             qDot should approximate(testGridQDotInfeed)
           case _ => fail("Thermal grid state updated failed")
         }
@@ -197,13 +231,15 @@ class ThermalGridWithStorageOnlySpec
               Some(
                 ThermalStorageState(
                   0L,
-                  KilowattHours(430d),
+                  KilowattHours(200d),
                   Kilowatts(0d),
                 )
               )
             ),
           testGridAmbientTemperature,
           testGridQDotConsumptionHigh,
+          thermalDemand,
+          noThermalDemand,
         ) match {
           case (
                 ThermalGridState(
@@ -213,7 +249,7 @@ class ThermalGridWithStorageOnlySpec
                 Some(StorageEmpty(thresholdTick)),
               ) =>
             tick shouldBe 0L
-            storedEnergy should approximate(KilowattHours(430d))
+            storedEnergy should approximate(KilowattHours(200d))
             qDot should approximate(testGridQDotConsumptionHigh)
             thresholdTick shouldBe 3600L
           case _ => fail("Thermal grid state updated failed")
@@ -226,6 +262,8 @@ class ThermalGridWithStorageOnlySpec
           ThermalGrid.startingState(thermalGrid),
           testGridAmbientTemperature,
           Kilowatts(0d),
+          noThermalDemand,
+          noThermalDemand,
         )
         updatedState match {
           case (
@@ -236,7 +274,7 @@ class ThermalGridWithStorageOnlySpec
                 None,
               ) =>
             tick shouldBe 0L
-            storedEnergy should approximate(KilowattHours(230d))
+            storedEnergy should approximate(KilowattHours(0d))
             qDot should approximate(Megawatts(0d))
 
           case _ => fail("Thermal grid state updated failed")